Separate lots of codes as library
This may help us make encapsulation and document generation better.
